# Improve performance of diagnostics and highlights