## Code changes

- Use `.operation()` for setting CodeMirror5 marks; should also help with porting to CM6 (where `RangeSets` will take over some of the responsibility)
- deprecate `highlight_range` and `highlight_markers` protected properties.

## User-facing changes

Improved performance when more than one diagnostic of highlighted element present in editor; especially noticeable in notebooks as the cost of relayout there is much higher, but also noticeable in file editor for large scripts.

### Highlights
- one cell with 1000 `math.pi` and `import math`; move cursor to `math`, wait for 1000 highlights, then move to `pi`:
  - before:
     - highlight `math`: 13.1s
     - then highlight `pi`: 16.6s
  - after:
      - highlight `math`: 160ms
       then highlight `pi`: 227ms
- 100 cells with `math.pi` and one with `import math`; move cursor to `math`, wait for 1000 highlights, then move to `pi` (this is overhead control, no gains expected):
   - before:
       - highlight `math`: 385ms
       - then highlight `pi`: 683 ms
    - after:
       - highlight `math`: 390ms
      - then highlight `pi`: 870ms

### Diagnostics

- one cell with 1000 `math.pi` and `import math`; comment out import, wait for 1000 diagnostics, then uncomment import, wait for removal:
  - before:
     - diagnostics show up in: 13.6s (blocking!)
     - diagnostics removal in: 13.2s (blocking!)
  - after:
     - diagnostics show up in: 254ms
     - diagnostics removal in: 160.4ms
- first open of a notebook with 10 cells, each with 88 diagnostics:
  - before: 2.75s (each time)
  - after 208.75ms
- first open of a notebook with 100 cells, each with 1 diagnostic this scenario is expected to have no gain (measures overhead)
  - before 280.34ms, 377ms, 399ms
  - after 385.29ms, 301.97ms, 309.4ms 

## Backwards-incompatible changes

None.

## Chores

- [x] linted <!-- Required: Run "jlpm lint" and "python scripts/lint.py" from the root of the repository, then check this box like this: [x] -->
- [ ] tested <!-- Recommended: Let us know if you already added a test case (if relevant). -->
- [ ] documented <!-- Optional: Would it be good to improve the documentation? If yes, please consider doing this and checking this box. -->
- [ ] changelog entry <!-- Recommended: Add a note in the CHANGELOG.md file under the most recent >unreleased< version; if one does not exist, feel free to create one by increasing the version number (no worries if you are not certain of the details - we can improve it later; let's just have something to work with) -->